reverted changes made in revision 1.10: fixed compiler warnings physics doesn't work quite right when using double in ua_physics_check_point_in_triangle()

+++ uwadv/source/underw/physics.cpp
@@ -619,7 +619,7 @@ bool ua_physics_check_point_in_triangle(const ua_vector3d& point,
    ua_vector3d e10=pb-pa;
    ua_vector3d e20=pc-pa;
      
-   double a,b,c,ac_bb;
+   float a,b,c,ac_bb;
 
    a = e10.dot(e10);
    b = e10.dot(e20);
@@ -628,12 +628,12 @@ bool ua_physics_check_point_in_triangle(const ua_vector3d& point,
 
    ua_vector3d vp = ua_vector3d(point.x-pa.x, point.y-pa.y, point.z-pa.z);
 
-   double d = vp.dot(e10);
-   double e = vp.dot(e20);
+   float d = vp.dot(e10);
+   float e = vp.dot(e20);
 
-   double x = (d*c)-(e*b);
-   double y = (e*a)-(d*b);
-   double z = x+y-ac_bb;
+   float x = (d*c)-(e*b);
+   float y = (e*a)-(d*b);
+   float z = x+y-ac_bb;
 
 #define in(a) ((Uint32&) a)
